Add GET /contacts/{id}/message endpoint
